refactor(nxz): extract parseMemlimitSize to its own module
The unit-test file imported `parseMemlimitSize` from `nxz.ts`, which
runs `main()` unconditionally on import. Under stricter vitest
configurations the module's `process.exit(1)` fires during test
collection.

Extract `parseMemlimitSize` into `packages/nxz/src/memlimit.ts` as a
side-effect-free helper module. `nxz.ts` imports + re-exports it for
public API stability. Test file points at the new module directly.

No behavior change: 29/29 unit cases still pass; 63 total in nxz
(2 test files); workspace test suite stays at 707 passing.
